pyproject.toml
[project]
name = "respondeai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "opensearch-py>=3.1.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.1",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "ipython>=9.6.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.1",
]
